chore: add better error message for ValueError
closes #1240
